backend/pyproject.toml
[tool.poetry]
name = "scheduler-backend"
version = "0.1.0"
description = "HighTech Plant scheduler backend"
authors = ["Anton <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
Django = ">=5.0.0,<5.1.0"
djangorestframework = ">=3.14.0"
django-cors-headers = ">=4.3.0"
django-filter= ">=24.2"
django-redis = ">=5.4"
drf-spectacular = ">=0.26.0"
Jinja2 = ">=3.1"
psycopg2-binary = ">=2.9.0"
pandas = ">=2.1.0"
openpyxl = ">=3.1.0"
python-docx = ">=1.1.0"
python-json-logger = ">=2.0.7"
regex = ">=2023.10.0"
rapidfuzz = ">=3.5.0"
dateparser = ">=1.1.0"
pydantic = ">=2.5.0"
requests = ">=2.31.0"
httpx = ">=0.25.0"
aiohttp = ">=3.9.0"
langchain = ">=0.1.0"
langchain-core = ">=0.1.0"
langchain-openai = ">=0.0.5"
langchain-community = ">=0.0.13"
tenacity = ">=8.2.0"
celery = ">=5.3.0"
redis = ">=5.0.0"
python-decouple = ">=3.8"
python-magic = ">=0.4.27"
pillow = ">=10.1.0"
pytz = ">=2023.3"
gunicorn = ">=21.2.0"
uvicorn = ">=0.24.0"
whitenoise = ">=6.6.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-django = ">=4.7.0"
pytest-asyncio = ">=0.21.0"
pytest-mock = ">=3.12.0"
factory-boy = ">=3.3.0"
hypothesis = ">=6.88.0"
black = ">=23.11.0"
ruff = ">=0.1.6"
mypy = ">=1.7.0"
pre-commit = ">=3.5.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
